Update script.js
Added a brick wall surrounding the play area, so that the player doesn't feel awkward that bullets dont go there
Ghost spawning is now better, and(hopefully) has no more bugs, and shows no more errors)
Ghosts can now move! However, currently they are only limited to following the player if the player is on the same X or Y coord as them. It logs "player dead" if the player is colliding with the ghoul.
Bullets now actually kill Ghouls
Abstracted a lot of stuff to functions so that it can be reuse with other enemies.