Update script.js
Added a new sleepy ghoul. This ghoul is fast, but will only attack if you are on the same axis as it.
Added doors for travel. Currently, only 1 test world exists to go right
coming back from a world to the main spawn area correctly puts you in the right corner, in accordance to the gate you went through.
Added flowers for aesthetic.
Refactoring