Update script.js
Added a new level and made changes to inversions. Encountered a very weird bug. Now sure what caused it, but basically, when i changes the player's x to the inverted's x, and same for y, the player coords were not being set on *2* blocks in the second level of the inverters... Due to this, the second level had to be redesigned. The problem still exists! In level 2, if you go to coords 3,5 and in inverted mode and chagne back to normal mode, your coords will not change.