Update script.js
FINISHED THE FINAL BOSS FIGHT!(almost)
Now, the boss has 4 phases, here is a breakdown of each one
In phase 1, the boss shoots one laser OR spawns an enemy.
In Phase 2, the amount of nuclear barrels is reduced. The boss shoots 2 lasers  OR spawns 2 enemies.
In phase 3, the amount of nuclear barrels is reduced. The boss shoots 3 laser OR spawns 2 enemies
In phase 4, the amount of nuclear barrels is reduced. 
In every subsequent phase, the boss attacks faster.
The boss is VERY hard to beat without all the gains of the other levels so good progression ig