Update script.js
Completed the game loop by finally adding an end screen. The game is now technically finished, all that's left is some polishing and QoL.
In the next commits, I will add music and an indicator of all your gems.
I also need to add a title screen and a game over screen.